+++ dev/com.ibm.ws.jsp/resources/com/ibm/ws/jsp/resources/messages.nlsprops
@@ -1263,10 +1263,10 @@ jsp.feature.not.loaded.correctly=CWWJS0001: The JSP [Pages] feature did not load
 jsp.feature.not.loaded.correctly.explanation=The JSP [Pages] feature is loaded with a version that corresponds to the Java [Jakarta] EE specification of that JSP [Pages] feature. The feature was loaded with no version defined.
 jsp.feature.not.loaded.correctly.useraction=Check the server.xml file for the existence of multiple JSP [Pages] features. Only one should be specified.
 
-jsp.error.page.invalid.erroronelnotfound=CWWJS0002E: Invalid value for errorOnELNotFound attribute
+jsp.error.page.invalid.erroronelnotfound=CWWJS0002: Invalid value for errorOnELNotFound attribute
 jsp.error.page.invalid.erroronelnotfound.explanation=
 jsp.error.page.invalid.erroronelnotfound.useraction=
 
-jsp.isthreadsafe.warning==CWWJS0003W: Per the Pages 3.1 specification, pages developers are strongly advised not to use the deprecated page directive isThreadSafe. 
+jsp.isthreadsafe.warning=CWWJS0003: Per the Pages 3.1 specification, pages developers are strongly advised not to use the deprecated page directive isThreadSafe. 
 jsp.isthreadsafe.warning.explanation=SingleThreadModel has been removed from Servlet 6.0. Pages 3.1 has deprecated isThreadSafe with plans to remove it in Page 4.0.
-jsp.isthreadsafe.warning.useraction=Remove the isThreadSafe directive from all pages.  
+jsp.isthreadsafe.warning.useraction=Remove the isThreadSafe directive from all pages.
