+++ dev/io.openliberty.pages.3.1.internal.factories/src/com/ibm/ws/pages31/translator/visitor/validator/ELValidatorExtImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.jsp23.translator.visitor.validator;
+package com.ibm.ws.pages31.translator.visitor.validator;
 
 import jakarta.el.ELException;
 import jakarta.el.ExpressionFactory;
