+++ dev/io.openliberty.pages.3.1.internal.factories/src/com/ibm/ws/pages31/translator/visitor/validator/ElValidatorExtFactoryImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,21 +8,21 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.jsp23.translator.visitor.generator;
+package com.ibm.ws.pages31.translator.visitor.validator;
 
 import org.osgi.service.component.annotations.Component;
 
-import com.ibm.ws.jsp.translator.visitor.generator.GeneratorUtilsExt;
-import com.ibm.ws.jsp.translator.visitor.generator.GeneratorUtilsExtFactory;
+import com.ibm.ws.jsp.translator.visitor.validator.ElValidatorExt;
+import com.ibm.ws.jsp.translator.visitor.validator.ElValidatorExtFactory;
 
 @Component(property = { "service.vendor=IBM" })
-public class GeneratorUtilsExtFactoryImpl implements GeneratorUtilsExtFactory {
+public class ElValidatorExtFactoryImpl implements ElValidatorExtFactory {
 
-    private static final GeneratorUtilsExtImpl gue = new GeneratorUtilsExtImpl();
+    private static final ELValidatorExtImpl eve = new ELValidatorExtImpl();
 
     @Override
-    public GeneratorUtilsExt getGeneratorUtilsExt() {
-        return gue;
+    public ElValidatorExt getELValidatorExt() {
+        return eve;
     }
 
 }
