+++ dev/io.openliberty.pages.3.1.internal.factories/src/com/ibm/ws/pages31/webcontainerext/ws/PrepareJspHelper23.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.jsp23.webcontainerext.ws;
+package com.ibm.ws.pages31.webcontainerext.ws;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayOutputStream;
