+++ dev/io.openliberty.pages.3.1.internal_fat/fat/src/io/openliberty/pages31/fat/FATSuite.java
@@ -31,7 +31,7 @@ import componenttest.rules.repeater.RepeatTests;
  */
 @RunWith(Suite.class)
 @SuiteClasses({
-    // TagsTestsEE10.class,
+    TagsTestsEE10.class,
     UnknownELIdentifiersTest.class
 })
 public class FATSuite {
